backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Validation & Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Export
pandas==2.2.0
openpyxl==3.1.2
